+++ packages/search-ui/src/__tests__/SearchDriver.test.js
@@ -1,47 +1,26 @@
 import SearchDriver, { DEFAULT_STATE } from "../SearchDriver";
 
-const mockApiConnector = {
-  search: jest.fn().mockReturnValue({ then: cb => cb(resultList) }),
-  click: jest.fn().mockReturnValue(Promise.resolve())
-};
+import { searchResponseWithoutFacets } from "../test/fixtures";
+import {
+  doesStateHaveResponseData,
+  setupDriver,
+  getMockApiConnector
+} from "../test/helpers";
 
-const resultList = {
-  info: {
-    facets: {},
-    meta: {
-      page: {
-        total_pages: 100,
-        total_results: 1000
-      },
-      request_id: "12345"
-    }
-  },
-  results: [{}, {}]
-};
-
-const resultListWithoutFacets = {
-  info: {
-    meta: {
-      page: {
-        total_pages: 100,
-        total_results: 1000
-      },
-      request_id: "12345"
-    }
-  },
-  results: [{}, {}]
-};
+const mockApiConnector = getMockApiConnector();
 
 const params = {
   apiConnector: mockApiConnector,
   trackUrlState: false
 };
 
+function getSearchCalls(specificMockApiConnector) {
+  return (specificMockApiConnector || mockApiConnector).search.mock.calls;
+}
+
 beforeEach(() => {
-  mockApiConnector.search = jest
-    .fn()
-    .mockReturnValue({ then: cb => cb(resultList) });
-  mockApiConnector.click = jest.fn().mockReturnValue({ then: () => {} });
+  mockApiConnector.search.mockClear();
+  mockApiConnector.click.mockClear();
 });
 
 it("can be initialized", () => {
@@ -54,79 +33,6 @@ it("will throw when missing required parameters", () => {
     new SearchDriver({});
   }).toThrow();
 });
-
-function setupDriver({ initialState, mockSearchResponse } = {}) {
-  const driver = new SearchDriver({
-    ...params,
-    initialState
-  });
-
-  if (mockSearchResponse) {
-    mockApiConnector.search = jest
-      .fn()
-      .mockReturnValue({ then: cb => cb(resultListWithoutFacets) });
-  }
-
-  const updatedStateAfterAction = {};
-  driver.subscribeToStateChanges(newState => {
-    updatedStateAfterAction.state = newState;
-  });
-
-  return {
-    stateAfterCreation: driver.getState(),
-    driver,
-    updatedStateAfterAction
-  };
-}
-
-function doesStateHaveResponseData(response) {
-  const {
-    requestId,
-    results,
-    totalPages,
-    totalResults,
-    wasSearched
-  } = response;
-  return (
-    !!results &&
-    !!requestId &&
-    results.length &&
-    totalPages > 0 &&
-    totalResults > 0 &&
-    !!wasSearched
-  );
-}
-
-function itResetsCurrent(fn) {
-  const state = fn();
-  it("resets current", () => {
-    expect(state.current).toEqual(1);
-  });
-}
-
-function itResetsFilters(fn) {
-  const state = fn();
-  it("resets filters", () => {
-    expect(state.filters).toEqual([]);
-  });
-}
-
-function itFetchesResults(fn) {
-  it("fetches results", () => {
-    const state = fn();
-    expect(doesStateHaveResponseData(state)).toBe(true);
-  });
-}
-
-function getSearchCalls() {
-  return mockApiConnector.search.mock.calls;
-}
-
-function waitABit(length) {
-  return new Promise(function(resolve) {
-    setTimeout(() => resolve(), length);
-  });
-}
 
 it("will use initial state if provided", () => {
   const initialState = {
@@ -151,7 +57,7 @@ it("will default facets to {} in state if facets is missing from the response", 
 
   const { stateAfterCreation } = setupDriver({
     initialState,
-    resultListWithoutFacets
+    searchResponseWithoutFacets
   });
 
   expect(doesStateHaveResponseData(stateAfterCreation)).toBe(true);
@@ -284,539 +190,5 @@ describe("#getActions", () => {
     expect(actions.setSort).toBeInstanceOf(Function);
     expect(actions.setCurrent).toBeInstanceOf(Function);
     expect(actions.trackClickThrough).toBeInstanceOf(Function);
-  });
-});
-
-describe("#setSearchTerm", () => {
-  function subject(term, { refresh, initialState = {} } = {}) {
-    const { driver, stateAfterCreation, updatedStateAfterAction } = setupDriver(
-      { initialState }
-    );
-    driver.setSearchTerm(term, { refresh });
-    return {
-      state: updatedStateAfterAction.state,
-      stateAfterCreation: stateAfterCreation
-    };
-  }
-
-  it("Updates searchTerm in state", () => {
-    expect(subject("test").state.searchTerm).toEqual("test");
-  });
-
-  it("Updates resultSearchTerm in state", () => {
-    expect(subject("test").state.resultSearchTerm).toEqual("test");
-  });
-
-  itResetsCurrent(
-    () => subject("test", { initialState: { current: 2 } }).state
-  );
-
-  itResetsFilters(
-    () =>
-      subject("test", { initialState: { filters: [{ filter1: ["value1"] }] } })
-        .state
-  );
-
-  it("Does not update other Search Parameter values", () => {
-    const initialState = {
-      resultsPerPage: 60,
-      sortField: "name",
-      sortDirection: "asc"
-    };
-    const { resultsPerPage, sortField, sortDirection } = subject("test", {
-      initialState
-    }).state;
-
-    expect({ resultsPerPage, sortField, sortDirection }).toEqual(initialState);
-  });
-
-  itFetchesResults(() => subject("term").state);
-
-  it("Does not fetch results when 'refresh' is set to false", () => {
-    expect(subject("term", { refresh: false }).state.wasSearched).toBe(false);
-  });
-
-  it("Will not debounce requests if there is no debounce specified", async () => {
-    const { driver } = setupDriver();
-    driver.setSearchTerm("term", { refresh: true });
-    driver.setSearchTerm("term", { refresh: true });
-    driver.setSearchTerm("term", { refresh: true });
-    expect(getSearchCalls()).toHaveLength(3);
-  });
-
-  it("Will debounce requests", async () => {
-    const { driver } = setupDriver();
-    driver.setSearchTerm("term", { refresh: true, debounce: 10 });
-    driver.setSearchTerm("term", { refresh: true, debounce: 10 });
-    driver.setSearchTerm("term", { refresh: true, debounce: 10 });
-    await waitABit(100);
-    expect(getSearchCalls()).toHaveLength(1);
-  });
-});
-
-describe("#addFilter", () => {
-  function subject(
-    name,
-    value,
-    {
-      initialFilters = [],
-      initialState = {
-        filters: initialFilters
-      }
-    } = {}
-  ) {
-    const { driver, updatedStateAfterAction } = setupDriver({ initialState });
-
-    driver.addFilter(name, value);
-    return updatedStateAfterAction.state;
-  }
-
-  itFetchesResults(() => subject("field", "value"));
-
-  itResetsCurrent(() =>
-    subject("field", "value", { initialState: { current: 2 } })
-  );
-
-  it("Does not update other Search Parameter values", () => {
-    const initialState = {
-      resultsPerPage: 60,
-      sortField: "name",
-      sortDirection: "asc",
-      searchTerm: "test"
-    };
-    const { resultsPerPage, sortField, sortDirection, searchTerm } = subject(
-      "field",
-      "value",
-      { initialState }
-    );
-    expect({ resultsPerPage, sortField, sortDirection, searchTerm }).toEqual(
-      initialState
-    );
-  });
-
-  it("Adds a new filter", () => {
-    expect(
-      subject("test", "value", { initialFilters: [{ initial: ["value"] }] })
-        .filters
-    ).toEqual([{ initial: ["value"] }, { test: ["value"] }]);
-  });
-
-  it("Adds an additional filter", () => {
-    expect(
-      subject("test", "value2", {
-        initialFilters: [{ initial: ["value"] }, { test: ["value"] }]
-      }).filters
-    ).toEqual([{ initial: ["value"] }, { test: ["value", "value2"] }]);
-  });
-
-  it("Won't add a duplicate filter", () => {
-    expect(
-      subject("test", "value", {
-        initialFilters: [{ initial: ["value"] }, { test: ["value"] }]
-      }).filters
-    ).toEqual([{ initial: ["value"] }, { test: ["value"] }]);
-  });
-
-  it("Supports range filters", () => {
-    expect(
-      subject(
-        "test",
-        {
-          from: 20,
-          to: 100
-        },
-        { initialFilters: [{ initial: ["value"] }] }
-      ).filters
-    ).toEqual([{ initial: ["value"] }, { test: [{ from: 20, to: 100 }] }]);
-  });
-
-  it("Adds an additional range filter", () => {
-    expect(
-      subject(
-        "test",
-        { from: 5, to: 6 },
-        {
-          initialFilters: [
-            { initial: [{ from: 20, to: 100 }] },
-            { test: [{ from: 4, to: 5 }] }
-          ]
-        }
-      ).filters
-    ).toEqual([
-      { initial: [{ from: 20, to: 100 }] },
-      { test: [{ from: 4, to: 5 }, { from: 5, to: 6 }] }
-    ]);
-  });
-
-  it("Won't add a duplicate range filter", () => {
-    expect(
-      subject(
-        "test",
-        {
-          from: 20,
-          to: 100
-        },
-        {
-          initialFilters: [
-            { initial: ["value"] },
-            { test: [{ from: 20, to: 100 }] }
-          ]
-        }
-      ).filters
-    ).toEqual([{ initial: ["value"] }, { test: [{ from: 20, to: 100 }] }]);
-  });
-});
-
-describe("#setFilter", () => {
-  function subject(
-    name,
-    value,
-    {
-      initialFilters = [],
-      initialState = {
-        filters: initialFilters
-      }
-    } = {}
-  ) {
-    const { driver, updatedStateAfterAction } = setupDriver({ initialState });
-
-    driver.setFilter(name, value);
-    return updatedStateAfterAction.state;
-  }
-
-  itFetchesResults(() => subject("field", "value"));
-
-  itResetsCurrent(() =>
-    subject("field", "value", { initialState: { current: 2 } })
-  );
-
-  it("Does not update other Search Parameter values", () => {
-    const initialState = {
-      resultsPerPage: 60,
-      sortField: "name",
-      sortDirection: "asc",
-      searchTerm: "test"
-    };
-    const { resultsPerPage, sortField, sortDirection, searchTerm } = subject(
-      "field",
-      "value",
-      { initialState }
-    );
-    expect({ resultsPerPage, sortField, sortDirection, searchTerm }).toEqual(
-      initialState
-    );
-  });
-
-  it("Adds a new filter and removes old filters", () => {
-    expect(
-      subject("test", "value2", { initialFilters: [{ initial: ["value"] }] })
-        .filters
-    ).toEqual([{ initial: ["value"] }, { test: ["value2"] }]);
-  });
-});
-
-describe("#removeFilter", () => {
-  function subject(
-    name,
-    value,
-    {
-      initialFilters = [],
-      initialState = {
-        filters: initialFilters
-      }
-    } = {}
-  ) {
-    const { driver, updatedStateAfterAction } = setupDriver({ initialState });
-
-    driver.removeFilter(name, value);
-    return updatedStateAfterAction.state;
-  }
-
-  itFetchesResults(() => subject("field", "value"));
-
-  itResetsCurrent(() =>
-    subject("field", "value", { initialState: { current: 2 } })
-  );
-
-  it("Does not update other Search Parameter values", () => {
-    const initialState = {
-      resultsPerPage: 60,
-      sortField: "name",
-      sortDirection: "asc",
-      searchTerm: "test"
-    };
-    const { resultsPerPage, sortField, sortDirection, searchTerm } = subject(
-      "field",
-      "value",
-      { initialState }
-    );
-    expect({ resultsPerPage, sortField, sortDirection, searchTerm }).toEqual(
-      initialState
-    );
-  });
-
-  it("Removes just 1 filter value", () => {
-    expect(
-      subject("test", "value", {
-        initialFilters: [
-          { initial: ["value"] },
-          { test: ["anotherValue", "value", "someOtherValue"] }
-        ]
-      }).filters
-    ).toEqual([
-      { initial: ["value"] },
-      { test: ["anotherValue", "someOtherValue"] }
-    ]);
-  });
-
-  it("Removes all filters", () => {
-    expect(
-      subject("test", undefined, {
-        initialFilters: [
-          { initial: ["value"] },
-          { test: ["anotherValue", "value", "someOtherValue"] }
-        ]
-      }).filters
-    ).toEqual([{ initial: ["value"] }]);
-  });
-
-  it("Removes all filters when last value", () => {
-    expect(
-      subject("test", "value", {
-        initialFilters: [{ initial: ["value"] }, { test: ["value"] }]
-      }).filters
-    ).toEqual([{ initial: ["value"] }]);
-  });
-
-  it("Removes just 1 range filter value", () => {
-    expect(
-      subject(
-        "test",
-        {
-          from: 20,
-          to: 100
-        },
-        {
-          initialFilters: [
-            { initial: [{ from: 20, to: 100 }] },
-            { test: ["anotherValue", { from: 20, to: 100 }, "someOtherValue"] }
-          ]
-        }
-      ).filters
-    ).toEqual([
-      { initial: [{ from: 20, to: 100 }] },
-      { test: ["anotherValue", "someOtherValue"] }
-    ]);
-  });
-});
-
-describe("#reset", () => {
-  it("Resets state back to the initial state provided at initialization", () => {
-    const initialState = {
-      resultsPerPage: 60,
-      sortField: "name",
-      sortDirection: "asc"
-    };
-
-    const { driver, updatedStateAfterAction } = setupDriver({ initialState });
-
-    driver.setSearchTerm("test");
-    let updatedStated = updatedStateAfterAction.state;
-
-    expect(updatedStated).not.toEqual({
-      ...DEFAULT_STATE,
-      ...initialState
-    });
-
-    driver.reset();
-    updatedStated = updatedStateAfterAction.state;
-
-    expect(updatedStated).toEqual({
-      ...DEFAULT_STATE,
-      ...initialState
-    });
-  });
-});
-
-describe("#clearFilters", () => {
-  function subject(
-    except,
-    {
-      initialFilters = [],
-      initialState = {
-        filters: initialFilters
-      }
-    } = {}
-  ) {
-    const { driver, updatedStateAfterAction } = setupDriver({ initialState });
-
-    driver.clearFilters(except);
-    return updatedStateAfterAction.state;
-  }
-
-  itResetsCurrent(() => subject(null, { initialState: { current: 2 } }));
-
-  it("Does not update other Search Parameter values", () => {
-    const initialState = {
-      resultsPerPage: 60,
-      sortField: "name",
-      sortDirection: "asc",
-      searchTerm: "test"
-    };
-    const { resultsPerPage, sortField, sortDirection, searchTerm } = subject(
-      null,
-      { initialState }
-    );
-    expect({ resultsPerPage, sortField, sortDirection, searchTerm }).toEqual(
-      initialState
-    );
-  });
-
-  it("Removes all filters", () => {
-    expect(
-      subject([], {
-        initialFilters: [
-          ({ initial: ["value"] },
-          { test: ["anotherValue", "value", "someOtherValue"] })
-        ]
-      }).filters
-    ).toEqual([]);
-  });
-
-  it("Removes all except the filters listed in 'except'", () => {
-    expect(
-      subject(["initial"], {
-        initialFilters: [
-          { initial: ["value"] },
-          { test: ["anotherValue", "value", "someOtherValue"] }
-        ]
-      }).filters
-    ).toEqual([{ initial: ["value"] }]);
-  });
-});
-
-describe("#setCurrent", () => {
-  function subject(current, { initialState = {} } = {}) {
-    const { driver, stateAfterCreation, updatedStateAfterAction } = setupDriver(
-      { initialState }
-    );
-    driver.setCurrent(current);
-    return {
-      state: updatedStateAfterAction.state,
-      stateAfterCreation: stateAfterCreation
-    };
-  }
-
-  it("Updates searchTerm in state", () => {
-    expect(
-      subject(2, { initialState: { searchTerm: "test" } }).state.current
-    ).toEqual(2);
-  });
-
-  it("Does not update other Search Parameter values", () => {
-    const initialState = {
-      searchTerm: "test",
-      filters: [{ initial: ["value"] }],
-      resultsPerPage: 60,
-      sortField: "name",
-      sortDirection: "asc"
-    };
-    const {
-      searchTerm,
-      filters,
-      resultsPerPage,
-      sortField,
-      sortDirection
-    } = subject(2, {
-      initialState
-    }).state;
-
-    expect({
-      searchTerm,
-      filters,
-      resultsPerPage,
-      sortField,
-      sortDirection
-    }).toEqual(initialState);
-  });
-});
-
-describe("#setSort", () => {
-  function subject(sortField, sortDirection, { initialState = {} } = {}) {
-    const { driver, stateAfterCreation, updatedStateAfterAction } = setupDriver(
-      { initialState }
-    );
-    driver.setSort(sortField, sortDirection);
-    return {
-      state: updatedStateAfterAction.state,
-      stateAfterCreation: stateAfterCreation
-    };
-  }
-
-  it("Updates sortField in state", () => {
-    expect(subject("date", "desc").state.sortField).toEqual("date");
-  });
-
-  it("Updates sortDirection in state", () => {
-    expect(subject("date", "desc").state.sortDirection).toEqual("desc");
-  });
-
-  itResetsCurrent(
-    () => subject("date", "desc", { initialState: { current: 2 } }).state
-  );
-
-  it("Does not update other Search Parameter values", () => {
-    const initialState = {
-      searchTerm: "test",
-      filters: [{ initial: ["value"] }],
-      resultsPerPage: 60
-    };
-    const { searchTerm, filters, resultsPerPage } = subject("date", "desc", {
-      initialState
-    }).state;
-
-    expect({
-      searchTerm,
-      filters,
-      resultsPerPage
-    }).toEqual(initialState);
-  });
-});
-
-describe("#setResultsPerPage", () => {
-  function subject(resultsPerPage, { initialState = {} } = {}) {
-    const { driver, stateAfterCreation, updatedStateAfterAction } = setupDriver(
-      { initialState }
-    );
-    driver.setResultsPerPage(resultsPerPage);
-    return {
-      state: updatedStateAfterAction.state,
-      stateAfterCreation: stateAfterCreation
-    };
-  }
-
-  it("Updates resultsPerPage in state", () => {
-    expect(subject(10).state.resultsPerPage).toEqual(10);
-  });
-
-  itResetsCurrent(() => subject(20, { initialState: { current: 2 } }).state);
-
-  it("Does not update other Search Parameter values", () => {
-    const initialState = {
-      searchTerm: "test",
-      filters: [{ initial: ["value"] }],
-      sortField: "date",
-      sortDirection: "desc"
-    };
-    const { searchTerm, filters, sortField, sortDirection } = subject(10, {
-      initialState
-    }).state;
-
-    expect({
-      searchTerm,
-      filters,
-      sortField,
-      sortDirection
-    }).toEqual(initialState);
   });
 });
