+++ packages/search-ui/src/__tests__/actions/setCurrent.test.js
@@ -1,0 +1,47 @@
+import { setupDriver } from "../../test/helpers";
+
+describe("#setCurrent", () => {
+  function subject(current, { initialState = {} } = {}) {
+    const { driver, stateAfterCreation, updatedStateAfterAction } = setupDriver(
+      { initialState }
+    );
+    driver.setCurrent(current);
+    return {
+      state: updatedStateAfterAction.state,
+      stateAfterCreation: stateAfterCreation
+    };
+  }
+
+  it("Updates searchTerm in state", () => {
+    expect(
+      subject(2, { initialState: { searchTerm: "test" } }).state.current
+    ).toEqual(2);
+  });
+
+  it("Does not update other Search Parameter values", () => {
+    const initialState = {
+      searchTerm: "test",
+      filters: [{ initial: ["value"] }],
+      resultsPerPage: 60,
+      sortField: "name",
+      sortDirection: "asc"
+    };
+    const {
+      searchTerm,
+      filters,
+      resultsPerPage,
+      sortField,
+      sortDirection
+    } = subject(2, {
+      initialState
+    }).state;
+
+    expect({
+      searchTerm,
+      filters,
+      resultsPerPage,
+      sortField,
+      sortDirection
+    }).toEqual(initialState);
+  });
+});
