+++ packages/search-ui/src/__tests__/actions/setFilter.test.js
@@ -1,0 +1,52 @@
+import { setupDriver } from "../../test/helpers";
+import { itResetsCurrent, itFetchesResults } from "../../test/sharedTests";
+
+describe("#setFilter", () => {
+  function subject(
+    name,
+    value,
+    {
+      initialFilters = [],
+      initialState = {
+        filters: initialFilters
+      }
+    } = {}
+  ) {
+    const { driver, updatedStateAfterAction } = setupDriver({
+      initialState
+    });
+
+    driver.setFilter(name, value);
+    return updatedStateAfterAction.state;
+  }
+
+  itFetchesResults(() => subject("field", "value"));
+
+  itResetsCurrent(() =>
+    subject("field", "value", { initialState: { current: 2 } })
+  );
+
+  it("Does not update other Search Parameter values", () => {
+    const initialState = {
+      resultsPerPage: 60,
+      sortField: "name",
+      sortDirection: "asc",
+      searchTerm: "test"
+    };
+    const { resultsPerPage, sortField, sortDirection, searchTerm } = subject(
+      "field",
+      "value",
+      { initialState }
+    );
+    expect({ resultsPerPage, sortField, sortDirection, searchTerm }).toEqual(
+      initialState
+    );
+  });
+
+  it("Adds a new filter and removes old filters", () => {
+    expect(
+      subject("test", "value2", { initialFilters: [{ initial: ["value"] }] })
+        .filters
+    ).toEqual([{ initial: ["value"] }, { test: ["value2"] }]);
+  });
+});
